Move FutureConverters to javaapi
